pyproject.toml
[project]
name = "rag1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
"streamlit>=1.25.0",
"openai>=1.3.5",
"langchain>=0.1.16",
"langchain-community>=0.0.32",
"langchain-openai>=0.0.6",
"qdrant-client>=1.7.0",
"python-dotenv>=1.0.0",
"PyPDF2>=3.0.1",
]
